Adding success/init callbacks and ignoreFormatting
For large amounts of data the plugin can take a long time to copy to the clipboard. I wanted to pop up a notification bubble in my app when the copy was done, but there was no way to do that with the basic plugin. So I went in and added `onCopyInit` and `onCopySuccess` callbacks, which trigger at the moment the user presses ctrl+C and the moment the large `for` loop that processes the data is done. It also exposes the number of rows that were copied.

I also had to add a way for the plugin to ignore formatters in selected columns -- for this I created the `ignoreFormatters` option, which is an array containing column field names. These columns get their formatting ignored when pasted, returing the raw JS. This is needed in cases where cells are formatted to HTML for browsers but we don't want to paste the HTML into Excel -- we want the value displayed.

Here's an example of initializing the plugin with these new options:

```javascript
this.externalCopyManager = new Slick.CellExternalCopyManager({
    onCopyInit: function() {
        console.log('starting copy');
    },
    onCopySuccess: function(rows) {
        console.log('Completed copying ' + rows + ' rows of data.');
    },
    overrideFormatting: ['field', 'anotherField']
});
```